unlock full power

+++ gen.cpp
@@ -80,7 +80,7 @@ int main(int argc, char ** argv) {
   double digits_per_iteration = 14.18;
 
   // precision is the amount of bits that are going to be used to represent the obtained pi: https://math.stackexchange.com/questions/160295/how-many-bits-needed-to-store-a-number
-  int precision = log2(10) *number_of_digits; 
+  long precision = log2(10) *number_of_digits; 
   int iterations = number_of_digits / digits_per_iteration + 1;
 
   mpf_set_default_prec(precision);
